requirements.txt
Python >= 3.7
PyTorch >= 1.7.0
pandas >= 1.0.0
numpy >= 1.18.0
scikit-learn >= 0.24.0
mamba-ssm >= 1.0.0
causal-conv1d >= 1.0.0